Introduces XProjectEnv (spitted IsolatedEnv as the base but without IDisposable)
+ Also adds an regular MSBuild Properties and other related changes for convenient work with this properties like extension method GetFileProperties, AddOrUpdate etc.

Part of https://github.com/3F/vsSolutionBuildEvent/pull/53
WIP
