+++ MvsSln/Core/IEnvironment.cs
@@ -24,110 +24,13 @@
 */
 
 using System;
-using System.Collections.Generic;
 using System.Runtime.InteropServices;
-using Microsoft.Build.Evaluation;
 
 namespace net.r_eg.MvsSln.Core
 {
     [Guid("818233C6-5BFE-47D5-929D-668C70EA25D5")]
-    public interface IEnvironment: IDisposable
+    public interface IEnvironment: IXProjectEnv, IDisposable
     {
-        /// <summary>
-        /// Access to Solution data.
-        /// </summary>
-        ISlnResult Sln { get; }
 
-        /// <summary>
-        /// List of all evaluated projects at current time 
-        /// with unique configuration for each instance.
-        /// </summary>
-        IEnumerable<IXProject> Projects { get; }
-
-        /// <summary>
-        /// List of evaluated projects that filtered by Guid.
-        /// </summary>
-        IEnumerable<IXProject> UniqueByGuidProjects { get; }
-
-        /// <summary>
-        /// Access to GlobalProjectCollection
-        /// </summary>
-        ProjectCollection PrjCollection { get; }
-
-        /// <summary>
-        /// Find project by Guid.
-        /// </summary>
-        /// <param name="guid">Guid of project.</param>
-        /// <param name="cfg">Specific configuration.</param>
-        /// <returns></returns>
-        IXProject XProjectByGuid(string guid, IConfPlatform cfg);
-
-        /// <summary>
-        /// Find project by Guid.
-        /// </summary>
-        /// <param name="guid">Guid of project.</param>
-        /// <returns></returns>
-        IXProject[] XProjectsByGuid(string guid);
-
-        /// <summary>
-        /// Find projects by name.
-        /// </summary>
-        /// <param name="name">ProjectName.</param>
-        /// <param name="cfg">Specific configuration.</param>
-        /// <returns></returns>
-        IXProject[] XProjectsByName(string name, IConfPlatform cfg);
-
-        /// <summary>
-        /// Find projects by name.
-        /// </summary>
-        /// <param name="name">ProjectName.</param>
-        /// <returns></returns>
-        IXProject[] XProjectsByName(string name);
-
-        /// <summary>
-        /// Get or firstly load into collection the project. 
-        /// Use default configuration.
-        /// </summary>
-        /// <param name="pItem">Specific project.</param>
-        /// <returns></returns>
-        Project GetOrLoadProject(ProjectItem pItem);
-
-        /// <summary>
-        /// Get or firstly load into collection the project.
-        /// </summary>
-        /// <param name="pItem">Specific project.</param>
-        /// <param name="conf">Configuration of project to load.</param>
-        /// <returns></returns>
-        Project GetOrLoadProject(ProjectItem pItem, IConfPlatform conf);
-
-        /// <summary>
-        /// Get or firstly load into collection the project.
-        /// </summary>
-        /// <param name="pItem">Specific project.</param>
-        /// <param name="properties"></param>
-        /// <returns></returns>
-        Project GetOrLoadProject(ProjectItem pItem, IDictionary<string, string> properties);
-
-        /// <summary>
-        /// Get project properties from solution properties.
-        /// </summary>
-        /// <param name="pItem"></param>
-        /// <param name="slnProps">Solution properties.</param>
-        /// <returns></returns>
-        IDictionary<string, string> GetProjectProperties(ProjectItem pItem, IDictionary<string, string> slnProps);
-
-        /// <summary>
-        /// Load available projects via configurations.
-        /// It will be added without unloading of previous.
-        /// </summary>
-        /// <param name="pItems">Specific list or null value to load all available.</param>
-        /// <returns>Loaded projects.</returns>
-        IEnumerable<IXProject> LoadProjects(IEnumerable<ProjectItemCfg> pItems = null);
-
-        /// <summary>
-        /// Load the only one configuration for each available project.
-        /// </summary>
-        /// <returns>Loaded projects.</returns>
-        IEnumerable<IXProject> LoadMinimalProjects();
     }
 }
